+++ packages/frontend/src/features/note/components/trash.tsx
@@ -31,7 +31,7 @@ const TrashItem = memo(function ({ noteId, className }: TrashItemProps) {
       tabIndex={0}
       onClick={() => navigateToNote(noteId)}
       className={cn(
-        "grid grid-cols-[24px_1fr_32px_32px] gap-1 items-center pr-1 pl-2 py-1 rounded-lg hover:bg-secondary/90 transition-colors duration-100",
+        "grid grid-cols-[24px_1fr_32px_32px] gap-1 items-center pr-1 pl-2 py-1 rounded-lg hover:bg-secondary/50 transition-colors duration-100",
         className,
       )}
     >
@@ -83,6 +83,7 @@ export function TrashWidget() {
       </PopoverTrigger>
       <PopoverContent
         side="right"
+        sticky="always"
         className="w-80 ml-2 grid grid-rows-[auto_1fr] bg-view-2/80 top-highlight max-h-[60vh] p-0 mb-2"
       >
         <div className="p-2">